+++ src/constants/ember-repos.js
@@ -1,7 +1,7 @@
-const adopted_ember_addons = require('./repos/adopted-ember-addons');
-const ember_cli = require('./repos/ember-cli');
-const ember_learn = require('./repos/ember-learn');
-const emberjs = require('./repos/emberjs');
+const adopted_ember_addons = require('./ember-organizations/adopted-ember-addons');
+const ember_cli = require('./ember-organizations/ember-cli');
+const ember_learn = require('./ember-organizations/ember-learn');
+const emberjs = require('./ember-organizations/emberjs');
 
 const community = [
   ...adopted_ember_addons,
